added Conn class, supertype of Client
