+++ mpmp-srv/src/main/Client.java
@@ -13,9 +13,7 @@ import cmds.Cmd;
  * chatting.
  * @author Leander, oki
  */
-public class Client {
-	private BufferedReader in;
-	private PrintWriter out;
+public class Client extends Conn {
 	private String name;
 
 	// XXX why is "private enum Mode {...} mode;" not allowed in Java? Because
@@ -27,49 +25,10 @@ public class Client {
 	private Mode mode;
 
 	public Client(BufferedReader in, PrintWriter out) {
-		this.in = in;
-		this.out = out;
+		super(in, out);
 		this.name = null;
 		this.mode = Mode.PreSubscribe;
 		send("Willkommen, Genosse! Subscriben Sie!");
-	}
-
-	/**
-	 * Handles the connection to a client, reading and writing commands and
-	 * replies.
-	 */
-	public void handle() throws SocketException, IOException {
-		String line, cmd;
-		Cmd c;
-		int delim;
-
-		for (;;) {
-			line = in.readLine();
-			delim = line.indexOf(' ');
-			if (delim < 0)
-				delim = line.length();
-
-			cmd = line.substring(0, delim);
-			c = Cmd.search(cmd);
-			if (c == null) {
-				sendErr("Command does not exist!");
-				continue;
-			}
-
-			c.exec(line, this);
-		}
-	}
-
-	public void send(String line) {
-		out.println(line);
-	}
-
-	public void sendOK() {
-		out.println("+JAWOHL");
-	}
-
-	public void sendErr(String s) {
-		out.println("-NEIN " + s);
 	}
 
 	public void subscribe(Mode mode, String name) {
